[turbopack] Always print trace labels in headers (#81728)
## Always format layer names in import trace headers

Previously if there was a single import trace where all items were in a single layer we wouldn't display the layer.  With this we always put the layer in the trace 'header' for a single trace just like we do when there are multiple traces (unless there are no layers which is really just a theoretical possibility).

### Why?

The concern was really about consistency

Tim was [confused](https://vercel.slack.com/archives/C03KAR5DCKC/p1752584756904199?thread_ts=1749510128.195809&cid=C03KAR5DCKC) why the layer headers showed up sometimes but not others.  Obviously when there are multiple traces, the layers are a critical distinguishing factor, when there is a single one we might suspect it is obvious, but inferring the layer of a file isn't always and by consistently formatting traces the same way we can ease readability.

Also add some unit tests to speed up iteration on this formatting.