Turbopack: Use workaround for rustc miscompilation bug on macos intel (#81950)
Uses https://github.com/hyperium/hyper/pull/3908 cherry-picked on top of
hyper v1.6.0 to work around
https://github.com/hyperium/hyper/issues/3902 and
https://github.com/rust-lang/rust/issues/140686

- Git Branch:
https://github.com/bgw/hyper-rs/commits/v1.6.0-with-macos-intel-miscompilation-workaround/
- Cherry-picked commit:
https://github.com/bgw/hyper-rs/commit/ab3544930722e6c270c16d3239dbb1d58f713393

Fixes #81697

The proper fix is to update the rustc toolchain, but since we use
nightly toolchains, this is too risky for a minor patch release. So this
applies the workaround instead.

## Test Plan

Force using x86-64 under Rosetta by downloading a "standalone" version
of nodejs and adding it to the front of the `PATH` with:

```
export PATH=/Users/bgw/Downloads/node-v22.17.1-darwin-x64/bin/:"$PATH"
```

Use `create-next-app` to create a new app. This fetching Geist by
default.

Run `pnpm dev`, see the 404 errors.

Modify the `package.json` to use this preview version:

```
    "next": "https://vercel-packages.vercel.app/next/commits/16196a3d2d7df65222727619e8db993336cf7aff/next"
```

Run `pnpm i && pnpm dev`, see that the 404s are solved!