Turbopack: Remove unused proxy option in turbo-tasks-fetch, lightly document HTTP_PROXY/HTTPS_PROXY environment variables (#81905)
We had this proxy option on `turbo-tasks-fetch`, but it was never passed anywhere.

I don't think the option is very useful, because `reqwest` already respects the conventional `HTTP_PROXY`/`HTTPS_PROXY` environment variables by default: https://docs.rs/reqwest/latest/reqwest/#proxies

It could make sense alongside the new experimental `turbopackUseSystemTlsCerts` option, but `turbopackUseSystemTlsCerts` is a simple boolean, and there's not a lot of harm in setting it to `true`, even if some of your users don't use system TLS certs. In contrast, HTTP proxies are very system-specific, and IMO make a lot more sense as an environment variable, often set in the user's `.profile` or in `/etc/environment`.

See the manual test plan on #81905 for how to set an HTTP/HTTPS proxy.