Optimize segment data routes (#82033)
This tweaks the segment data routes we create for dynamic routes to be
derived from the `__PAGE__` segment route as we can create one route to
handle the 3 segment outputs for a dynamic route.

This also removes the segment data routes for static routes as they are
not needed and be routed to without an explicit regex route per entry.