Turbopack: write tasks doesn't need to be session dependent, as effects will restore (#78727)
### What?

Filesystem write task don't need to be session dependent. We don't need to reexecute them for that reason. We already emit an "effect" and that is still emitted after restoring from cache.

Not making them session dependent avoids dirty updates and also tracking the task as dirty in the aggregated graph. 

But the "effect" is not-serializeable, so the write task will be recomputed to recompute the value of the effect, when trying to apply the effect.