[turbopack] Improve our `const` compatible hash routine performance (#82088)
The explicit operators can apparently not be folded by the compiler.

This is the const compatible way to take a slice and convert a slice to an array... sigh.

https://godbolt.org/z/Wedn7hWvb  shows the 3 possible implementations.

The `std` option appears worse than it really would in practice where the bounds checks could be eliminated, but pointer arithmetic gets the goods nonetheless.

Discovered this while trying to integrate my 'static `rcstr!`' fix into SWC.  It ended up failing for subtle reasons but the benchmarks did reveal a hash regression which we can at least fix on our side.