Revert "Fix tracing of server actions imported by client components (#78968) (#82161)
It seems these additional fs operations/transforms on client components
while tracing dependencies is causing a big perf hit which is not worth
the edge case of catching fs dependencies in server actions from client
components.

A workaround is available without our automatic tracing via
`outputFileTracingIncludes` so this reverts the tracing to alleviate the
perf concern.

We will be able to handle cases like this without the perf consequence
with turbopack going forward.

This reverts commit af0473b7f43bbabe0b327ec65348bf85610a1153.

Closes: https://github.com/vercel/next.js/issues/81902